more precise terrain painting brush

+++ src/studio/property_view/terrain_editor.h
@@ -20,6 +20,7 @@ class RenderScene;
 class Texture;
 }
 
+
 class TerrainEditor : public Lumix::WorldEditor::Plugin
 {
 	friend class TerrainComponentPlugin;
@@ -84,7 +85,7 @@ private:
 	Lumix::ComponentUID m_component;
 	QTreeWidgetItem* m_texture_tree_item;
 	float m_terrain_brush_strength;
-	int m_terrain_brush_size;
+	float m_terrain_brush_size;
 	int m_texture_idx;
 	QString m_selected_entity_template;
 };
